[feature/jquery_clock] Add jQuery-based clock to prosilver clock_side
The clock we currently use needs flash support. As newer browsers and
mobile devices are moving on from flash to HTML5 we should do the same and
replace our flash clock with a jQuery based clock.
It is using image sprites which means we only have to load 2 images. One
for the single digits and one for the double digit hours.
This is currently only added to prosilver and the obsolete settings in the
clock module still need to be removed.

B3P-218